refactor: split config and path ownership
Problem: polylogue.paths mixed directory layout, configuration loading, model types, and sanitization helpers behind one facade, while Config.db_path still resolved through a service-locator property.

What changed: turn polylogue.paths into a directory-layout package, move Source/DriveConfig/IndexConfig and configuration loaders into polylogue.config, require explicit polylogue.paths.sanitize imports for safety helpers, and document environment precedence.

Verification: devtools verify; focused config/path/render/backend tests passed during the branch.

Closes #191
