+++ polylogue/config.py
@@ -1,26 +1,80 @@
-"""Minimal runtime configuration derived from filesystem defaults and env overrides."""
+"""Runtime configuration derived from filesystem defaults and env overrides."""
 
 from __future__ import annotations
 
-from dataclasses import dataclass
+import os
+from dataclasses import dataclass, field
 from pathlib import Path
 
 from .errors import PolylogueError
 from .paths import (
-    DriveConfig,
-    IndexConfig,
-    Source,
+    GEMINI_DRIVE_FOLDER,
     archive_root,
-    db_path,
-    get_drive_config,
-    get_index_config,
-    get_sources,
+    claude_code_path,
+    codex_path,
+    drive_cache_path,
+    drive_credentials_path,
+    drive_token_path,
+    inbox_root,
     render_root,
+)
+from .paths import (
+    db_path as default_db_path,
 )
 
 
 class ConfigError(PolylogueError):
     """Configuration error."""
+
+
+@dataclass
+class Source:
+    """A conversation source (local path, Drive folder, or both)."""
+
+    name: str
+    path: Path | None = None
+    folder: str | None = None
+
+    def __post_init__(self) -> None:
+        if not self.name or not self.name.strip():
+            raise ValueError("Source name cannot be empty")
+        self.name = self.name.strip()
+        has_path = self.path is not None
+        has_folder = self.folder is not None and self.folder.strip()
+        if not has_path and not has_folder:
+            raise ValueError(f"Source '{self.name}' must have either 'path' or 'folder'")
+        if self.folder:
+            self.folder = self.folder.strip()
+
+    @property
+    def is_drive(self) -> bool:
+        return self.folder is not None
+
+
+@dataclass
+class DriveConfig:
+    """Google Drive OAuth configuration."""
+
+    credentials_path: Path = field(default_factory=drive_credentials_path)
+    token_path: Path = field(default_factory=drive_token_path)
+    retry_count: int = 3
+    timeout: int = 30
+
+
+@dataclass
+class IndexConfig:
+    """Search indexing configuration."""
+
+    fts_enabled: bool = True
+    voyage_api_key: str | None = None
+
+    @classmethod
+    def from_env(cls) -> IndexConfig:
+        """Load IndexConfig from environment variables."""
+        return cls(
+            fts_enabled=True,
+            voyage_api_key=os.environ.get("VOYAGE_API_KEY"),
+        )
 
 
 @dataclass
@@ -30,13 +84,41 @@ class Config:
     archive_root: Path
     render_root: Path
     sources: list[Source]
+    db_path: Path = field(default_factory=default_db_path)
     drive_config: DriveConfig | None = None
     index_config: IndexConfig | None = None
 
-    @property
-    def db_path(self) -> Path:
-        """Database path (always XDG default)."""
-        return db_path()
+
+def get_sources() -> list[Source]:
+    """Return the configured conversation sources."""
+    sources = [Source(name="inbox", path=inbox_root())]
+
+    if claude_code_path().exists():
+        sources.append(Source(name="claude-code", path=claude_code_path()))
+
+    if codex_path().exists():
+        sources.append(Source(name="codex", path=codex_path()))
+
+    gemini_cache = drive_cache_path() / "gemini"
+    if gemini_cache.exists() or drive_credentials_path().exists() or drive_token_path().exists():
+        sources.append(
+            Source(
+                name="gemini",
+                folder=GEMINI_DRIVE_FOLDER,
+                path=gemini_cache,
+            )
+        )
+    return sources
+
+
+def get_drive_config() -> DriveConfig:
+    """Return Drive configuration with default paths."""
+    return DriveConfig()
+
+
+def get_index_config() -> IndexConfig:
+    """Return index configuration from environment."""
+    return IndexConfig.from_env()
 
 
 def get_config() -> Config:
@@ -45,6 +127,7 @@ def get_config() -> Config:
         archive_root=archive_root(),
         render_root=render_root(),
         sources=get_sources(),
+        db_path=default_db_path(),
         drive_config=get_drive_config(),
         index_config=get_index_config(),
     )
@@ -57,4 +140,7 @@ __all__ = [
     "IndexConfig",
     "Source",
     "get_config",
+    "get_drive_config",
+    "get_index_config",
+    "get_sources",
 ]
