+++ polylogue/paths/_roots.py
