--- polylogue/paths_models.py
@@ -1,59 +1,0 @@
-"""Path-related data models."""
-
-from __future__ import annotations
-
-import os
-from dataclasses import dataclass, field
-from pathlib import Path
-
-from polylogue.paths_roots import drive_credentials_path, drive_token_path
-
-
-@dataclass
-class Source:
-    """A conversation source (local path, Drive folder, or both)."""
-
-    name: str
-    path: Path | None = None
-    folder: str | None = None
-
-    def __post_init__(self) -> None:
-        if not self.name or not self.name.strip():
-            raise ValueError("Source name cannot be empty")
-        self.name = self.name.strip()
-        has_path = self.path is not None
-        has_folder = self.folder is not None and self.folder.strip()
-        if not has_path and not has_folder:
-            raise ValueError(f"Source '{self.name}' must have either 'path' or 'folder'")
-        if self.folder:
-            self.folder = self.folder.strip()
-
-    @property
-    def is_drive(self) -> bool:
-        return self.folder is not None
-
-
-@dataclass
-class DriveConfig:
-    """Google Drive OAuth configuration."""
-
-    credentials_path: Path = field(default_factory=drive_credentials_path)
-    token_path: Path = field(default_factory=drive_token_path)
-    retry_count: int = 3
-    timeout: int = 30
-
-
-@dataclass
-class IndexConfig:
-    """Search indexing configuration."""
-
-    fts_enabled: bool = True
-    voyage_api_key: str | None = None
-
-    @classmethod
-    def from_env(cls) -> IndexConfig:
-        """Load IndexConfig from environment variables."""
-        return cls(
-            fts_enabled=True,
-            voyage_api_key=os.environ.get("VOYAGE_API_KEY"),
-        )
