+++ polylogue/rendering/renderers/markdown.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from pathlib import Path
 
-from polylogue.paths import conversation_render_root
+from polylogue.paths.sanitize import conversation_render_root
 from polylogue.rendering.core import ConversationFormatter
