+++ tests/integration/test_security.py
@@ -8,7 +8,7 @@ from pathlib import Path
 
 import pytest
 
-from polylogue.paths import Source
+from polylogue.config import Source
 from polylogue.sources.decoders import MAX_UNCOMPRESSED_SIZE
 from polylogue.sources.source_parsing import iter_source_conversations
 from polylogue.storage.cursor_state import CursorFailurePayload, CursorStatePayload
